BackEnd/requirements.txt
absl-py==2.3.0
adal==1.2.7
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiomysql==0.2.0
aiopg==1.4.0
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
APScheduler==3.11.1
asgiref==3.8.1
async-timeout==4.0.3
asyncio==4.0.0
asyncpg==0.30.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
av==13.0.0
azure-common==1.1.28
azure-core==1.35.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-nspkg==3.0.2
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-storage-blob==12.25.1
azure-storage-common==1.4.2
beautifulsoup4==4.13.4
billiard==4.2.1
boto3==1.40.7
botocore==1.40.7
celery==5.3.4
certifi==2025.8.3
cffi==1.17.1
channels==4.2.2
channels-redis==4.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
Cython==3.1.6
daphne==4.2.0
decorator==5.2.1
decouple==0.0.7
deep-translator==1.11.4
distro==1.9.0
Django==5.0.14
django-cors-headers==4.7.0
django-extensions==4.1
django-mssql-backend==2.8.1
django-redis==5.4.0
djangorestframework==3.16.0
dnspython==2.7.0
docx2txt==0.9
dotenv==0.9.9
duckduckgo_search==8.0.4
easydict==1.13
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.5.1
graphviz==0.21
groq==0.37.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
incremental==24.7.2
insightface==0.7.3
isodate==0.7.2
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.8
kombu==5.5.4
lazy_loader==0.4
livekit==1.0.20
livekit-api==1.1.0
livekit-protocol==1.1.1
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.3
moviepy==2.2.1
mpmath==1.3.0
msal==1.32.3
msgpack==1.1.1
msrest==0.7.1
msrestazure==0.6.4.post1
mssql-django==1.5
multidict==6.6.2
mysql==0.0.3
mysql-connector-python==9.4.0
mysqlclient==2.1.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
obsws-python==1.8.0
onnx==1.19.1
onnxruntime-gpu==1.18.0
openai==2.11.0
opencv-contrib-python==4.8.1.78
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
Pillow==10.1.0
platformdirs==4.5.0
prettytable==3.16.0
primp==0.15.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycuda==2025.1.2
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyee==11.1.1
PyJWT==2.8.0
pymongo==4.13.2
PyMySQL==1.1.1
pyodbc==5.2.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyppeteer==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytools==2025.2.5
pytz==2025.2
PyYAML==6.0.2
redis==7.1.0
regex==2025.10.23
reportlab==4.4.4
requests==2.32.4
requests-oauthlib==2.0.0
s3transfer==0.13.1
sacremoses==0.1.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
screeninfo==0.8.1
selenium==4.9.1
sentencepiece==0.2.1
service-identity==24.2.0
setproctitle==1.3.7
simsimd==6.5.3
siphash24==1.8
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.7
sqlparse==0.5.3
starlette==0.46.2
stringzilla==4.2.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.1
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.57.1
trio==0.30.0
trio-websocket==0.12.2
triton==3.3.1
Twisted==25.5.0
txaio==23.6.1
types-protobuf==6.30.2.20250703
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.159
ultralytics-thop==2.0.14
urllib3==1.26.18
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
whitenoise==6.6.0
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
